move helpers from internal to api package and expose utility function
