+++ backend/backend.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 
 	"github.com/dapr/durabletask-go/api"
-	"github.com/dapr/durabletask-go/internal/helpers"
+	"github.com/dapr/durabletask-go/api/helpers"
 	"github.com/dapr/durabletask-go/internal/protos"
 	"google.golang.org/protobuf/proto"
 )
