+++ tests/backend_test.go
@@ -9,9 +9,9 @@ import (
 	"time"
 
 	"github.com/dapr/durabletask-go/api"
+	"github.com/dapr/durabletask-go/api/helpers"
 	"github.com/dapr/durabletask-go/backend"
 	"github.com/dapr/durabletask-go/backend/sqlite"
-	"github.com/dapr/durabletask-go/internal/helpers"
 	"github.com/dapr/durabletask-go/internal/protos"
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/protobuf/types/known/timestamppb"
