+++ tests/task_executor_test.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/dapr/durabletask-go/api"
-	"github.com/dapr/durabletask-go/internal/helpers"
+	"github.com/dapr/durabletask-go/api/helpers"
 	"github.com/dapr/durabletask-go/internal/protos"
 	"github.com/dapr/durabletask-go/task"
 	"github.com/stretchr/testify/require"
