+++ tests/worker_test.go
@@ -7,8 +7,8 @@ import (
 	"time"
 
 	"github.com/dapr/durabletask-go/api"
+	"github.com/dapr/durabletask-go/api/helpers"
 	"github.com/dapr/durabletask-go/backend"
-	"github.com/dapr/durabletask-go/internal/helpers"
 	"github.com/dapr/durabletask-go/internal/protos"
 	"github.com/dapr/durabletask-go/tests/mocks"
 	"github.com/stretchr/testify/assert"
